Combine some interfaces
