+++ src/main/java/net/fabricmc/loom/configuration/CompileConfiguration.java
@@ -41,7 +41,6 @@ import net.fabricmc.loom.configuration.ide.SetupIntelijRunConfigs;
 import net.fabricmc.loom.configuration.providers.LaunchProvider;
 import net.fabricmc.loom.configuration.providers.MinecraftProviderImpl;
 import net.fabricmc.loom.configuration.providers.mappings.MappingsProviderImpl;
-import net.fabricmc.loom.extension.internal.ConfigurationInternalGradleExtension;
 import net.fabricmc.loom.util.Constants;
 
 public final class CompileConfiguration {
@@ -51,30 +50,30 @@ public final class CompileConfiguration {
 	public static void setupConfigurations(Project project) {
 		final ConfigurationContainer configurations = project.getConfigurations();
 
-		ConfigurationInternalGradleExtension data = LoomGradleExtension.get(project);
+		LoomGradleExtension extension = LoomGradleExtension.get(project);
 
-		data.createLazyConfiguration(Constants.Configurations.MOD_COMPILE_CLASSPATH).configure(configuration -> configuration.setTransitive(true));
-		data.createLazyConfiguration(Constants.Configurations.MOD_COMPILE_CLASSPATH_MAPPED).configure(configuration -> configuration.setTransitive(false));
-		data.createLazyConfiguration(Constants.Configurations.MINECRAFT_NAMED).configure(configuration -> configuration.setTransitive(false)); // The launchers do not recurse dependencies
-		data.createLazyConfiguration(Constants.Configurations.MINECRAFT_DEPENDENCIES).configure(configuration -> configuration.setTransitive(false));
-		data.createLazyConfiguration(Constants.Configurations.LOADER_DEPENDENCIES).configure(configuration -> configuration.setTransitive(false));
-		data.createLazyConfiguration(Constants.Configurations.MINECRAFT).configure(configuration -> configuration.setTransitive(false));
-		data.createLazyConfiguration(Constants.Configurations.INCLUDE).configure(configuration -> configuration.setTransitive(false)); // Dont get transitive deps
-		data.createLazyConfiguration(Constants.Configurations.MAPPING_CONSTANTS);
+		extension.createLazyConfiguration(Constants.Configurations.MOD_COMPILE_CLASSPATH).configure(configuration -> configuration.setTransitive(true));
+		extension.createLazyConfiguration(Constants.Configurations.MOD_COMPILE_CLASSPATH_MAPPED).configure(configuration -> configuration.setTransitive(false));
+		extension.createLazyConfiguration(Constants.Configurations.MINECRAFT_NAMED).configure(configuration -> configuration.setTransitive(false)); // The launchers do not recurse dependencies
+		extension.createLazyConfiguration(Constants.Configurations.MINECRAFT_DEPENDENCIES).configure(configuration -> configuration.setTransitive(false));
+		extension.createLazyConfiguration(Constants.Configurations.LOADER_DEPENDENCIES).configure(configuration -> configuration.setTransitive(false));
+		extension.createLazyConfiguration(Constants.Configurations.MINECRAFT).configure(configuration -> configuration.setTransitive(false));
+		extension.createLazyConfiguration(Constants.Configurations.INCLUDE).configure(configuration -> configuration.setTransitive(false)); // Dont get transitive deps
+		extension.createLazyConfiguration(Constants.Configurations.MAPPING_CONSTANTS);
 
 		extendsFrom(JavaPlugin.COMPILE_ONLY_CONFIGURATION_NAME, Constants.Configurations.MAPPING_CONSTANTS, project);
 
-		data.createLazyConfiguration(Constants.Configurations.MAPPINGS);
-		data.createLazyConfiguration(Constants.Configurations.MAPPINGS_FINAL);
-		data.createLazyConfiguration(Constants.Configurations.LOOM_DEVELOPMENT_DEPENDENCIES);
-		data.createLazyConfiguration(Constants.Configurations.UNPICK_CLASSPATH);
+		extension.createLazyConfiguration(Constants.Configurations.MAPPINGS);
+		extension.createLazyConfiguration(Constants.Configurations.MAPPINGS_FINAL);
+		extension.createLazyConfiguration(Constants.Configurations.LOOM_DEVELOPMENT_DEPENDENCIES);
+		extension.createLazyConfiguration(Constants.Configurations.UNPICK_CLASSPATH);
 
 		for (RemappedConfigurationEntry entry : Constants.MOD_COMPILE_ENTRIES) {
-			data.createLazyConfiguration(entry.sourceConfiguration())
+			extension.createLazyConfiguration(entry.sourceConfiguration())
 					.configure(configuration -> configuration.setTransitive(true));
 
 			// Don't get transitive deps of already remapped mods
-			data.createLazyConfiguration(entry.getRemappedConfiguration())
+			extension.createLazyConfiguration(entry.getRemappedConfiguration())
 					.configure(configuration -> configuration.setTransitive(false));
 
 			extendsFrom(entry.getTargetConfiguration(configurations), entry.getRemappedConfiguration(), project);
